requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==5.1.0
Django==3.1
django-filter==21.1
djangorestframework==3.12.4
importlib-metadata==4.8.2
ipython==7.16.2
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.3.6
model-mommy==2.0.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
typing-extensions==4.0.1
wcwidth==0.2.5
zipp==3.6.0
